Bugfix: Fix array bounds issue in Slice filter
